Initial commit
This project is going to grow up to be an experimental promise lib that
I'm creating for learning purposes.

However, this is a great starting point for javascript libraries that
will execute both in the browser and in node-land.

Chase McCarthy (@code0100fun) worked on a project recently that used
this structure.  I've removed the parts of his project that aren't
related to build process and created this one.

If you'd like to use this here's what you do

1. clone it down
2. replace `pinky-swear` anywhere it occurs in your cloned directory
with the name of your project
3. rm -rf the `.git` directory and re-init

This will give you the `npm test` command which will build your files
into the `dist` directory and wrap them in UMD format (which is a
universal loading mechanism) so that it will work in node/browser.

`npm test` will also launch testem and will live reload upon file
changes.  To test in the browser navigate to the URL that testem prints
at the top left-hand corner of the terminal

caveat emptor

Have fun!